% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bco_mbd_sim.R
\name{bco_mbd_sim}
\alias{bco_mbd_sim}
\title{Conformized name of \link{mbd_sim}}
\usage{
bco_mbd_sim(mbd_params, crown_age = NA, stem_age = NA,
  conditioned_on = "nothing")
}
\arguments{
\item{mbd_params}{MBD parameters,
as created by \link{create_mbd_params}}

\item{crown_age}{the crown age.
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{stem_age}{the stem age
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{conditioned_on}{conditioning of the simulation, can be:
\itemize{
  \item \code{nothing} no conditioning, species can all go extinct
  \item \code{no_extinction} species cannot all go extinct
}}
}
\description{
Conformized name of \link{mbd_sim}
}
\author{
Richel J.C. Bilderbeek
}
