% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bco_pbd_sim.R
\name{bco_pbd_sim}
\alias{bco_pbd_sim}
\title{Higher-level interface to \link{pbd_sim}}
\usage{
bco_pbd_sim(pbd_params, crown_age = NA, stem_age = NA,
  add_shortest_and_longest = FALSE)
}
\arguments{
\item{pbd_params}{PBD parameters,
as created by \link{create_pbd_params}}

\item{crown_age}{the crown age.
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{stem_age}{the stem age
Either \code{crown_age} or \code{stem_age} must be exclusively set}
}
\description{
Higher-level interface to \link{pbd_sim}
}
\author{
Richel J.C. Bilderbeek
}
