% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_sim.R
\name{bd_sim}
\alias{bd_sim}
\title{Simulate a constant-rate birth-death process.}
\usage{
bd_sim(bd_params, crown_age = NA, stem_age = NA,
  conditioned_on = "nothing")
}
\arguments{
\item{bd_params}{BD parameters,
as created by \link{create_bd_params}}

\item{crown_age}{the crown age.
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{stem_age}{the stem age
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{conditioned_on}{conditioning of the simulation, can be:
\itemize{
  \item \code{nothing} no conditioning, species can all go extinct
  \item \code{no_extinction} species cannot all go extinct
}}
}
\description{
Simulate a constant-rate birth-death process.
}
\author{
Richel J.C. Bilderbeek
}
