% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditioned_on.R
\name{check_conditioned_on}
\alias{check_conditioned_on}
\title{Check if the conditioning is a proper descriptor}
\usage{
check_conditioned_on(conditioned_on)
}
\arguments{
\item{conditioned_on}{what the process is conditioned on.
This can be either 'nothing' or 'non_extinction'"}
}
\description{
Will \link{stop} if not
}
