% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crown_age.R
\name{check_crown_age}
\alias{check_crown_age}
\title{Check if the crown age is valid}
\usage{
check_crown_age(crown_age)
}
\arguments{
\item{crown_age}{the crown age, in time units ago}
}
\description{
Will \link{stop} if not.
}
