% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_each_pbd_param_selected_once.R
\name{check_each_pbd_param_selected_once}
\alias{check_each_pbd_param_selected_once}
\title{Check if the two \code{pbd_params_selector}s
together select each PBD parameter once.}
\usage{
check_each_pbd_param_selected_once(fixed_params, opt_params)
}
\arguments{
\item{fixed_params}{first \code{pbd_params_selector}}

\item{opt_params}{second \code{pbd_params_selector}}
}
\description{
Will \link{stop} if not
}
