% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_init_n_species.R
\name{check_init_n_species}
\alias{check_init_n_species}
\title{Check if the initial number of species equals one or two}
\usage{
check_init_n_species(init_n_species)
}
\arguments{
\item{init_n_species}{initial number of species at the
start of simulating a phylogeny. Can be 1 (to start at the
stem) or 2 (to start at the crown)}
}
\description{
Will \link{stop} if not
}
