% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mbd_params_selector.R
\name{check_mbd_params_selector}
\alias{check_mbd_params_selector}
\title{Check if the supplied object in an MBD parameter selector.}
\usage{
check_mbd_params_selector(mbd_params_selector)
}
\arguments{
\item{mbd_params_selector}{a structure to select MBD parameters,
as can be created by \link{create_mbd_params_selector}}
}
\description{
Will link{stop} if not.
}
\examples{
library(testthat)
expect_silent(check_mbd_params_selector(create_mbd_params_selector()))
expect_error(check_mbd_params_selector("nonsense"))
}
\author{
Richel J.C. Bilderbeek
}
