% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pbd_params.R
\name{check_pbd_params}
\alias{check_pbd_params}
\title{Check a PBD parameters set.}
\usage{
check_pbd_params(pbd_params)
}
\value{
nothing.
}
\description{
Will \link{stop} if invalid, will do nothing otherwise
}
\author{
Richel J.C. Bilderbeek
}
