% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pbd_params_selector.R
\name{check_pbd_params_selector}
\alias{check_pbd_params_selector}
\title{Check if the input is a valid PBD parameter selector.}
\usage{
check_pbd_params_selector(pbd_params_selector)
}
\arguments{
\item{pbd_params_selector}{a structure to select PBD parameters,
as can be created by \link{create_pbd_params_selector}}
}
\description{
Will \link{stop} if not
}
\seealso{
use \link{is_pbd_params_selector} to measure
if the input is a valid PBD parameter selector
}
