% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stem_age.R
\name{check_stem_age}
\alias{check_stem_age}
\title{Check if the stem age is valid}
\usage{
check_stem_age(stem_age)
}
\arguments{
\item{stem_age}{the stem age, in time units ago}
}
\description{
Will \link{stop} if not.
}
