% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pbd_params.R
\name{create_pbd_params}
\alias{create_pbd_params}
\title{Create the PBD parameters}
\usage{
create_pbd_params(erg, eri, scr, sirg, siri)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{scr}{speciation completion rate}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}
}
\description{
Create the PBD parameters
}
\author{
Richel J.C. Bilderbeek
}
