% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pbd_params_selector.R
\name{create_pbd_params_selector}
\alias{create_pbd_params_selector}
\title{Create a PBD parameter selector}
\usage{
create_pbd_params_selector(
  erg = FALSE,
  eri = FALSE,
  scr = FALSE,
  sirg = FALSE,
  siri = FALSE
)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{scr}{speciation completion rate}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}
}
\description{
Create a PBD parameter selector
}
\author{
Richel J.C. Bilderbeek
}
