% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bd_param_names.R
\name{get_bd_param_names}
\alias{get_bd_param_names}
\title{Get the names of the parameters used in the BD model}
\usage{
get_bd_param_names()
}
\description{
Get the names of the parameters used in the BD model
}
\author{
Richel J.C. Bilderbeek
}
