% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbd_param_names.R
\name{get_pbd_param_names}
\alias{get_pbd_param_names}
\title{Get the names of the parameters used in the PBD model}
\usage{
get_pbd_param_names()
}
\description{
Get the names of the parameters used in the PBD model
}
\author{
Richel J.C. Bilderbeek
}
