% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_find_scenario.R
\name{get_pbd_scenario_function}
\alias{get_pbd_scenario_function}
\title{Gets a function to determine if a scenario is true}
\usage{
get_pbd_scenario_function(scenario_description)
}
\arguments{
\item{scenario_description}{the scenario description}
}
\description{
Gets a function to determine if a scenario is true
}
