% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bd_params.R
\name{is_bd_params}
\alias{is_bd_params}
\title{Is object \code{x} an BD parameter set,
  as can be created by \code{create_bd_params}?}
\usage{
is_bd_params(x)
}
\arguments{
\item{x}{the object to be determined of if
its an BD parameter set}
}
\value{
TRUE if yes, else FALSE
}
\description{
Is object \code{x} an BD parameter set,
  as can be created by \code{create_bd_params}?
}
\author{
Richel J.C. Bilderbeek
}
