% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mbd_params_selector.R
\name{is_mbd_params_selector}
\alias{is_mbd_params_selector}
\title{Determine if the supplied object in an MBD parameter selector,
as can be created by \code{create_mbd_params_selector}}
\usage{
is_mbd_params_selector(mbd_params_selector)
}
\arguments{
\item{mbd_params_selector}{a structure to select MBD parameters,
as can be created by \link{create_mbd_params_selector}}
}
\value{
TRUE if x is an MBD parameters selector
}
\description{
Determine if the supplied object in an MBD parameter selector,
as can be created by \code{create_mbd_params_selector}
}
\examples{
library(testthat)

s <- create_mbd_params_selector()

expect_true(is_mbd_params_selector(s))
expect_false(is_mbd_params_selector("nonsense"))
}
\author{
Richel J.C. Bilderbeek
}
