% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pbd_params.R
\name{is_pbd_params}
\alias{is_pbd_params}
\title{Is object \code{x} an PBD parameter set,
  as can be created by \code{create_pbd_params}?}
\usage{
is_pbd_params(x)
}
\arguments{
\item{x}{the object to be determined of if
its an PBD parameter set}
}
\value{
TRUE if yes, else FALSE
}
\description{
Is object \code{x} an PBD parameter set,
  as can be created by \code{create_pbd_params}?
}
\author{
Richel J.C. Bilderbeek
}
