% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pbd_params_selector.R
\name{is_pbd_params_selector}
\alias{is_pbd_params_selector}
\title{Determine if the supplied object in an PBD parameter selector,
as can be created by \code{create_pbd_params_selector}}
\usage{
is_pbd_params_selector(x)
}
\arguments{
\item{x}{the object to be determined if it is an PBD
parameters selector}
}
\value{
TRUE if x is an PBD parameters selector
}
\description{
Determine if the supplied object in an PBD parameter selector,
as can be created by \code{create_pbd_params_selector}
}
\examples{
s <- create_pbd_params_selector()

library(testthat)
expect_true(is_pbd_params_selector(s))
expect_false(is_pbd_params_selector("nonsense"))
}
\author{
Richel J.C. Bilderbeek
}
