% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_calc_max_lik.R
\name{pbd_calc_max_lik}
\alias{pbd_calc_max_lik}
\title{Do an PBD maximum likelihood estimate.}
\usage{
pbd_calc_max_lik(branching_times, init_param_values, fixed_params,
  opt_params, init_n_species = 2, n_missing_species = 0,
  conditioned_on = "nothing")
}
\arguments{
\item{branching_times}{the branching times of the phylogeny}

\item{init_param_values}{initial parameter values,
as created with \code{create_pbd_params}}

\item{fixed_params}{the parameters that are fixed.
This is a list of the five parameter names, and a
boolean to indicate if it is fixed yes/no.
Use \link{create_pbd_params_selector} to create
such a list}

\item{opt_params}{the parameters that are estimated.
This is a list of the five parameter names, and a
boolean to indicate if it is optimized yes/no.
Use \link{create_pbd_params_selector} to create
such a list}

\item{init_n_species}{the number of species at the moment of the
first branching time. Can be either one or two:
\itemize{
  \item 1: stem
  \item 2: crown
}}

\item{n_missing_species}{the number of species missing}

\item{conditioned_on}{on what must be the likelihood estimation be
conditioned:
\itemize{
  \item \code{"nothing"}: the branch lengths were obtained from
    a random phylogeny that could also have gone extinct
  \item \code{"non_extinction"}: the branch lengths were obtained from
    a phylogeny conditioned on non-extinction
}}
}
\description{
Do an PBD maximum likelihood estimate.
}
\author{
Richel J.C. Bilderbeek
}
