% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec_checked.R
\name{pbd_numspec_median_checked}
\alias{pbd_numspec_median_checked}
\title{Calcates the median number of species under protracted birth-death model of
diversification. This is a checked version with a more explicit interface
than \code{\link{pbd_numspec_median}}.}
\usage{
pbd_numspec_median_checked(erg, eri, scr, sir, crown_age)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{scr}{speciation completion rate}

\item{sir}{speciation initiation rate. This function assumes the
speciation initiation rates of good and incipient species are equal.}

\item{crown_age}{the crown age}
}
\value{
The median number of representative species
}
\description{
Calcates the median number of species under protracted birth-death model of
diversification. This is a checked version with a more explicit interface
than \code{\link{pbd_numspec_median}}.
}
\examples{
  median_n_species <- pbd_numspec_median_checked(
    erg = 1.2,
    eri = 2.3,
    scr = 3.4,
    sir = 4.5,
    crown_age = 6.7
  )
  median_n_species_expected <- 99408712
  testthat::expect_equal(median_n_species, median_n_species_expected)
}
\seealso{
\code{\link{pbd_numspec_median}} provides for setting a time-dependence
  in the parameters and/or specify a stem age.
  \code{\link{pbd_numspec_mean_checked}} calculates the mean number of species.
}
\author{
Richel J.C. Bilderbeek
}
