% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bco_pbd_sim.R
\name{bco_pbd_sim}
\alias{bco_pbd_sim}
\title{Higher-level interface to \link{pbd_sim}}
\usage{
bco_pbd_sim(
  pbd_params,
  crown_age = NULL,
  stem_age = NULL,
  add_shortest_and_longest = FALSE
)
}
\arguments{
\item{pbd_params}{PBD parameters,
as created by \link{create_pbd_params}}

\item{crown_age}{the crown age.
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{stem_age}{the stem age
Either \code{crown_age} or \code{stem_age} must be exclusively set}

\item{add_shortest_and_longest}{if TRUE, the result of this
function will also include the species trees obtained
by sampling with 'shortest' and 'longest'}
}
\description{
Higher-level interface to \link{pbd_sim}
}
\author{
Richel J.C. Bilderbeek
}
