% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_init_n_species.R
\name{check_init_n_species}
\alias{check_init_n_species}
\title{Check if the initial number of species equals one or two}
\usage{
check_init_n_species(init_n_species)
}
\description{
Will \link{stop} if not
}
