% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mbd_params.R
\name{check_mbd_params}
\alias{check_mbd_params}
\title{Check a MBD parameters set.}
\usage{
check_mbd_params(mbd_params)
}
\arguments{
\item{mbd_params}{parameter set for the MBD model,
as can be created by \link{create_mbd_params}}
}
\value{
nothing.
}
\description{
Will \link{stop} if invalid, will do nothing otherwise
}
\author{
Richel J.C. Bilderbeek
}
