% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_params.R
\name{create_bd_params}
\alias{create_bd_params}
\title{Create the BD parameters}
\usage{
create_bd_params(ext_rate, spec_rate)
}
\description{
Create the BD parameters
}
\author{
Richel J.C. Bilderbeek
}
