% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(crown_age, erg, eri, ext_rate, lambda, mbd_params, mu,
  nu, pbd_params, pbd_params_selector, q, quantile, scr, sir, sirg, siri,
  spec_rate, stem_age)
}
\arguments{
\item{crown_age}{the crown age, in time units ago}

\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{ext_rate}{per-lineage extinction rate}

\item{lambda}{the sympatric speciation rate}

\item{mbd_params}{parameter set for the MBD model,
as can be created by \link{create_mbd_params}}

\item{mu}{the extinction rate}

\item{nu}{the multiple allopatric speciation trigger rate}

\item{pbd_params}{parameter set for the PBD model,
as can be created by \link{create_pbd_params}}

\item{pbd_params_selector}{a structure to select PBD parameters,
as can be created by \link{create_pbd_params_selector}}

\item{q}{the single-lineage speciation probability at a triggered event}

\item{quantile}{a quantile, a value between (and including) zero
to (and including) one.}

\item{scr}{speciation completion rate}

\item{sir}{speciation initiation rate}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{spec_rate}{per-lineage speciation rate}

\item{stem_age}{the stem age, in time units ago}

\item{mbd_params_selector}{a structure to select MBD parameters,
as can be created by \link{create_mbd_params_selector}}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
