% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(crown_age, erg, eri, lambda, mu, nu, q, quantile, scr,
  sir, sirg, siri, stem_age)
}
\arguments{
\item{crown_age}{the crown age, in time units ago}

\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{lambda}{the sympatric speciation rate}

\item{mu}{the extinction rate}

\item{nu}{the multiple allopatric speciation trigger rate}

\item{q}{the single-lineage speciation probability at a triggered event}

\item{quantile}{a quantile, a value between (and including) zero
to (and including) one.}

\item{scr}{speciation completion rate}

\item{sir}{speciation initiation rate}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{stem_age}{the stem age, in time units ago}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
